requirements.txt
django==3.1.2
djangorestframework==3.7.7
psycopg2==2.8.6
django-cors-headers==3.5.0